Support new jiri-swift generated building of cgo library & headers.
Use the unified generated header and rename the shared types.h to a more explicit
go_types.h.

Also bitcode is turned off as CGO cannot generate bitcode currently (it does not
use LLVM).

Does not fully build yet as we have illegal text-relocation errors in the linker.
This needs to be addressed in a future CL.

MultiPart: 2/2
Change-Id: Ib0c1348009f7b530cb377d80b0d333754625af2b