SyncbaseCore: Link to snappy, leveldb, and libc++
Adds snappy, leveldb to the Xcode project and adds -libc++ to the
linker flags.

The binary that CGO generates does not statically link these used
libraries, and it is a good thing since we might want to offer the
lib without these automatically linked so that we don't conflict
with the end-user's existing lib usage.

MultiPart: 2/2
Change-Id: Id83923699774b45572da20c105e3e3bf941edbf1