swift: Use wildcard permissions/blessings to fix unit tests
The blessing enforcement that was added by v.io/c/22770 and
v.io/c/22771 broke the existing swift Syncbase unit tests. This CL
fixes this break by using the '...' as a blessing and explicitly
passing permissions that similarly use the '...' wildcard.

This CL mirrors the JNI fixes in v.io/c/22962

Change-Id: Id4f3ba1204993b54474e5b8846a04ba0a3f58b2b