UX / UI only: Invite flow implementation
Shows a segmented list of nearby contacts, nearby, and contacts, with
find-as-you-type search. If no matches are found during search, a
unique cell is shown allowing for an invite to be sent to an email.
The email is regex validated and an alert is shown if invalid. The
secondary line of device listing for each person was omitted because
it was unclear which UX mock should be used.

Change-Id: I1b81b94bb0b5a59820ec37f5089c085e7a448d32