java: Get the DatabaseOptions.rootDir to work
Note that the rootDir is not exposed in the io.v.syncbase.core. The
purpose of this CL is to get all the high-level API tests from
SyncbaseTest to work. This change also adds the ability to shutdown
syncbase. The high-level API tests are reconfigured to use and empty
list of mounts points speed up the shutdown.

MultiPart: 3/3
Change-Id: I142966b1a2128d0b3de136d41fb0504f254ba672