swift: Create high-level Syncbase framework
Initial implementation of the Sycnbase high level API and framework.
Builds on top of the SyncbaseCore framework and its CGO library so
no changes yet needed for the jiri swift command. This CL achieves
parity with the current Java high-level API, although it has not
been tested much. Unit tests and bug fixes to come.

Note unit tests are currently not passing due to blessings issues
that are being worked on wrt to init and login.

This CL has following additional changes to the SyncbaseCore
framework to make this high-level work possible:

- Match the HLAPI in terms of configure/init/login, pass rootDir
  to v23_syncbase_Init()
- Remove Threads.swift as its largely unused.
- Implement Database.syncgroup, Database.listSyncgroups,
- Collection.exists(row)
- Make public many things that were incorrectly internal
- Create public initializers for structs
- Add runInBatchSync support function + unit test for Syncbase
  high-level
- Add Exist/NoExist errors to SyncbaseError
- Create a queue public variable to let the user control threads for
  callbacks
- Make ResumeMarker a typealias instead of a struct
- Refactors the OAuth credentials to not force the Google Credential
  in SyncbaseCore, uses the Syncbase.queue for the callback
- Consistently refers to UTF-8 instead of UTF8 in comments and
  error descriptions, matching the HLAPI.
- Make sure callbacks are always on Syncbase.queue and not main.

Change-Id: I82c765d269486a7e6c332f796a6896a0ef107677