syncbase: Get the login to work by adding a testLogin option
This change adds a testLogin option to the set of option we can pass
to Init in the high-level Syncbase API. When set to true the testLogin
will cause the call to login() to generate a 'root:o:app:user'
blessing which will set the app name to 'app' and the user name to
'user'. This also fixes the tests from internal/BlessingsTest.

Note: the author of the Swift code is Aaron Zinman <aaron@azinman.com>.

MultiPart: 3/3
Change-Id: I74acd8c6c68bf2c46c640e98a425b650161c8e42