swift: Fixes to enable dice roller to work, plus logout
Until merge is working, uses a scheme whereby each device has its
own unique collection name but updates the UI on any dice collection
change. While each device has its own collection, the UI looks as if
it has just the one, thereby achieving the desired UX.

NOTE: This will quickly change as we'll be using the userdata
collection instead, but that work will come in another CL. This is
more of a transition CL.

Logout is implemented for dice roller.

This CL contains the following other fixes for this example to work:

- Make User a class & hashable so it may be stored in a Set/Map & used
  as a segue sender (via AnyObject cast)
- Allow stopScan to work even if not yet init/logged in
- LoginCallback uses SyncbaseError instead of ErrorType
- Syncbase.configure's default for rootDir is the corrected
  defaultRootDir variable
- Added NoAccess exception, used for knowing when to delete the
  Syncbase rootDir after logging out and back in as a different user
- Rename the typo defaultSyncbasePerms -> defaultSyncgroupPerms

Change-Id: Ia2856eb8426ce86ffb27f1eb8f41b9e325371980