swift: Use UUID-based names when creating collections; bug fixes
To avoid unexpected collisions until syncgroup merge is implemented,
we generate a random UUID for each collection, prefaced by an optional
prefix. The API change is as follows:

Database.collection(name) -> Database.createCollection(prefix)

This matches the work in Java in CL 23722.

It also exposes the userdata collection in general, hides the internal
syncgroup id storage in it from watches on userdata, and updates the
dice roller example to use the userdata collection.

This CL also has the following other bug fixes:

- Implements a hack for canceling a watch until #1392 is fixed
- Improves error reporting in unit tests when an exception is thrown
- Fixes deserialization of AccessLists
- Implements debugDescription for SyncgroupSpec

Change-Id: I476b2ae6170cd5ec785b3b8273fc3464fd748aa8