swift: Fix persisted login for dice roller & todos
Now that credentials are being persisted in CGO, checks in the
Login controllers for isLoggedIn during viewDidLoad were attempting
to segue right away. Unfortuantely that cannot be done in viewDidLoad
so the segue get's scheduled for later execution on main.

Also replaces the custom activity indicator in Dice Roller with the
system network activity indicator, thereby also fixing the invalid
constraints in the storyboard.

Change-Id: I9787cd834efab7742f6829932dc9a5865c4431a3