swift: Update todos app with working model
Takes UX shell and plumbs it with Syncbase and a live data model.
Uses a central watchpoint and replicates the data in-memory, dispatching
notifications to screens of changes. Also saves the Google Sign In URL
for photos. This will be a useful hook once we figure out a strategy
for querying all member photos or advertising such urls in discovery.

Invite is not yet implemented, but syncing across a given user's
device is working (at least from a Swift perspective, various errors
in joining syncgroups and crashes are occuring in the Go layer).

Other fixes:

- Allow hashValue computation of Identifier to overflow rather than
  panic
- Disable bitcode on todos

Change-Id: Iafc8d5418661a39fa7abb14e8abe43c1801cfa34